+++ CHANGELOG.md
@@ -56,12 +56,14 @@ If upgrading from v2.x, see the [v3.0.0 release notes](https://github.com/flixOp
 If upgrading from v2.x, see the [v3.0.0 release notes](https://github.com/flixOpt/flixOpt/releases/tag/v3.0.0) and [Migration Guide](https://flixopt.github.io/flixopt/latest/user-guide/migration-guide-v3/).
 
 ### ✨ Added
+- exporting logger.SUCESS level
 
 - Added proper deprecation tests
 
 ### 💥 Breaking Changes
 
 ### ♻️ Changed
+- logger coloring improved
 
 ### 🗑️ Deprecated